Apply TransactionController metric builder pattern
