+++ app/core/Engine/controllers/transaction-controller/constants.ts
@@ -1,0 +1,6 @@
+import type { TransactionMetrics } from './types';
+
+export const EMPTY_METRICS: TransactionMetrics = Object.freeze({
+  properties: {},
+  sensitiveProperties: {},
+});
