+++ app/core/Engine/controllers/transaction-controller/types.ts
@@ -1,10 +1,12 @@
-import { JsonMap } from '../../../Analytics/MetaMetrics.types';
+import {
+  JsonMap,
+  IMetaMetricsEvent,
+} from '../../../Analytics/MetaMetrics.types';
 import { SmartTransactionsController } from '@metamask/smart-transactions-controller';
 import type { RootState } from '../../../../reducers';
 import { TransactionControllerInitMessenger } from '../../messengers/transaction-controller-messenger';
 import { TransactionMeta } from '@metamask/transaction-controller';
 
-// In order to not import from redux slice, type is defined here
 export interface TransactionMetrics {
   properties: JsonMap;
   sensitiveProperties: JsonMap;
@@ -16,9 +18,15 @@ export interface TransactionEventHandlerRequest {
   smartTransactionsController: SmartTransactionsController;
 }
 
-export type TransactionMetricsBuilder = (request: {
+export interface TransactionMetricsBuilderRequest {
+  eventType: IMetaMetricsEvent;
   transactionMeta: TransactionMeta;
   allTransactions: TransactionMeta[];
   getUIMetrics: (transactionId: string) => TransactionMetrics;
   getState: () => RootState;
-}) => TransactionMetrics;
+  transactionEventHandlerRequest: TransactionEventHandlerRequest;
+}
+
+export type TransactionMetricsBuilder = (
+  request: TransactionMetricsBuilderRequest,
+) => TransactionMetrics | Promise<TransactionMetrics>;
